Fixed invalid casting to int, thanks [DeCo]
git-svn-id: http://svn.l2jserver.com/trunk@4004 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/com/l2jserver/gameserver/model/actor/instance/L2PcInstance.java
@@ -5955,7 +5955,7 @@ public final class L2PcInstance extends L2Playable
 		if (getExpBeforeDeath() > 0)
 		{
 			// Restore the specified % of lost experience.
-			getStat().addExp((int)Math.round((getExpBeforeDeath() - getExp()) * restorePercent / 100));
+			getStat().addExp(Math.round((getExpBeforeDeath() - getExp()) * restorePercent / 100));
 			setExpBeforeDeath(0);
 		}
 	}
